# new: Improve CLI startup time and performance.