--- a/packages/cli/react.d.ts
+++ b/packages/cli/react.d.ts
@@ -1,0 +1,1 @@
+export * from './dts/react';