--- a/packages/cli/src/components/Help.tsx
+++ b/packages/cli/src/components/Help.tsx
@@ -1,30 +1,32 @@
-/* eslint-disable promise/prefer-await-to-then */
-
-import React, { Suspense } from 'react';
+import React from 'react';
 import { Box, Text } from 'ink';
 import { OptionConfigMap, ParamConfigList } from '@boost/args';
 import { SPACING_COL } from '../constants';
 import { formatDescription } from '../helpers';
 import { Categories, CommandConfig, CommandConfigMap } from '../types';
 import { Header } from './Header';
+import { HelpCommands } from './internal/HelpCommands';
+import { HelpOptions } from './internal/HelpOptions';
+import { HelpParams } from './internal/HelpParams';
+import { HelpUsage } from './internal/HelpUsage';
 
-function extractDefault<K extends string, P>(
-	key: K,
-): (result: Record<K, React.ComponentType<P>>) => { default: React.ComponentType<P> } {
-	return ({ [key]: component }) => ({ default: component });
-}
+// function extractDefault<K extends string, P>(
+// 	key: K,
+// ): (result: Record<K, React.ComponentType<P>>) => { default: React.ComponentType<P> } {
+// 	return ({ [key]: component }) => ({ default: component });
+// }
 
-const Commands = React.lazy(() =>
-	import('./internal/HelpCommands').then(extractDefault('HelpCommands')),
-);
+// const Commands = React.lazy(() =>
+// 	import('./internal/HelpCommands').then(extractDefault('HelpCommands')),
+// );
 
-const Options = React.lazy(() =>
-	import('./internal/HelpOptions').then(extractDefault('HelpOptions')),
-);
+// const Options = React.lazy(() =>
+// 	import('./internal/HelpOptions').then(extractDefault('HelpOptions')),
+// );
 
-const Params = React.lazy(() => import('./internal/HelpParams').then(extractDefault('HelpParams')));
+// const Params = React.lazy(() => import('./internal/HelpParams').then(extractDefault('HelpParams')));
 
-const Usage = React.lazy(() => import('./internal/HelpUsage').then(extractDefault('HelpUsage')));
+// const Usage = React.lazy(() => import('./internal/HelpUsage').then(extractDefault('HelpUsage')));
 
 export interface HelpProps {
 	categories?: Categories;
@@ -62,15 +64,13 @@ export function Help({
 				</Box>
 			)}
 
-			<Suspense fallback={null}>
-				{hasUsage && <Usage delimiter={delimiter} usage={config?.usage} />}
+			{hasUsage && <HelpUsage delimiter={delimiter} usage={config?.usage} />}
 
-				{hasParams && <Params config={config} params={params!} />}
+			{hasParams && <HelpParams config={config} params={params!} />}
 
-				{hasCommands && <Commands categories={categories} commands={commands!} />}
+			{hasCommands && <HelpCommands categories={categories} commands={commands!} />}
 
-				{hasOptions && <Options categories={categories} options={options!} />}
-			</Suspense>
+			{hasOptions && <HelpOptions categories={categories} options={options!} />}
 		</Box>
 	);
 }