--- a/packages/cli/src/helpers/applyMarkdown.ts
+++ b/packages/cli/src/helpers/applyMarkdown.ts
@@ -5,9 +5,11 @@ export function applyMarkdown(text: string): string {
 		if (match.startsWith('~')) {
 			return style.strikethrough(body);
 		}
+
 		if (match.startsWith('**') || match.startsWith('__')) {
 			return style.bold(body);
 		}
+
 		if (match.startsWith('*') || match.startsWith('_')) {
 			return style.italic(body);
 		}