I've noticed that programs built on Boost take about 1.5 seconds on average before _actually_ outputting anything to the console... and also taking 3 seconds in total for something as simple as the help menu to render... This is bad...

I originally thought the problem was either a) the react/ink rendering cycle is slow, b) args parsing and middleware is slow, or c) running the program in general is slow. Turns out, it was none of these. React rendering is about 250ms at the highest, while args parsing is about 10ms, and running the program is dependent on rendering, so average at about 250ms also. So where is the performance bottleneck???

Turns out that simply`require`ing and evaluating all of the Boost files is the primary culprit. Just loading Boost libs take about 1.5 seconds, with the CLI package accounting for almost 1s worth. This is... unfortunate. To mitigate this problem, I am doing the following:

- Defer importing of React components until they actually used. This helps avoid the `require` load and the evaluation hit of the component (since components make up half of the package itself).
  - In v3, all of the React stuff will be moved behind `@boost/cli/react`.
  - Since this is now async, we can parallelize component rendering a bit.
- ~Utilized `React.lazy` and `Suspense` for rendering improvements.~ [Blocked by unit tests](https://github.com/vadimdemedes/ink-testing-library/issues/18)
- Split components down into function components, so that they can be memoized more effectively.
  
 In the end, I was able to improve performance by almost 50%! 

- CLI `require` load: 1.1s -> 750ms (600ms with React code moved to a separate file).
- Time to first output: 1.5s avg -> 1s avg
- Time to complete: 3s avg -> 1.7s avg